requirements.txt
pyamiibo==0.2
requests==2.22.0
nfcpy==1.0.3
